Make TensorFlow dependencies optional
