+++ model_card_toolkit/__init__.py
@@ -13,7 +13,39 @@
 # limitations under the License.
 """A module to streamline and automate generation of Model Cards."""
 from model_card_toolkit.core import ModelCardToolkit
-from model_card_toolkit.model_card import *
+from model_card_toolkit.model_card import (
+    Citation, ConfidenceInterval, Considerations, Dataset, Graphic,
+    GraphicsCollection, KeyVal, License, Limitation, ModelCard, ModelDetails,
+    ModelParameters, Owner, PerformanceMetric, QuantitativeAnalysis, Reference,
+    Risk, SensitiveData, Tradeoff, UseCase, User, Version
+)
 from model_card_toolkit.tfx.component import ModelCardGenerator
-from model_card_toolkit.utils.source import *
 from model_card_toolkit.version import __version__
+
+__all__ = [
+    '__version__',
+    'Citation',
+    'ConfidenceInterval',
+    'Considerations',
+    'Dataset',
+    'Graphic',
+    'GraphicsCollection',
+    'KeyVal',
+    'License',
+    'Limitation',
+    'ModelCard',
+    'ModelCardGenerator',
+    'ModelCardToolkit',
+    'ModelDetails',
+    'ModelParameters',
+    'Owner',
+    'PerformanceMetric',
+    'QuantitativeAnalysis',
+    'Reference',
+    'Risk',
+    'SensitiveData',
+    'Tradeoff',
+    'UseCase',
+    'User',
+    'Version',
+]
