+++ model_card_toolkit/base_model_card_field.py
@@ -24,7 +24,7 @@ from typing import Any, Dict
 
 from google.protobuf import descriptor, message
 
-from model_card_toolkit.utils import validation
+from model_card_toolkit.utils import json_utils
 
 
 class BaseModelCardField(abc.ABC):
@@ -102,8 +102,9 @@ class BaseModelCardField(abc.ABC):
           setattr(self, field_name, [])
           for p in getattr(proto, field_name):
             # To get the type hint of a list is not easy.
-            field = \
-              self.__annotations__[field_name].__args__[0]()  # pytype: disable=attribute-error
+            field = (
+                self.__annotations__[field_name].__args__[0]()  # pytype: disable=attribute-error
+            )
             field._from_proto(p)  # pylint: disable=protected-access
             getattr(self, field_name).append(field)
 
@@ -136,7 +137,7 @@ class BaseModelCardField(abc.ABC):
   ) -> "BaseModelCardField":
     """Parses a JSON dictionary into the current object."""
     for subfield_key, subfield_json_value in json_dict.items():
-      if subfield_key.startswith(validation.SCHEMA_VERSION_STRING):
+      if subfield_key.startswith(json_utils.SCHEMA_VERSION_STRING):
         continue
       elif not hasattr(field, subfield_key):
         raise ValueError(
@@ -151,8 +152,9 @@ class BaseModelCardField(abc.ABC):
         subfield_value = []
         for item in subfield_json_value:
           if isinstance(item, dict):
-            new_object = \
-              field.__annotations__[subfield_key].__args__[0]()  # pytype: disable=attribute-error
+            new_object = (
+                field.__annotations__[subfield_key].__args__[0]()  # pytype: disable=attribute-error
+            )
             subfield_value.append(self._from_json(item, new_object))
           else:  # if primitive
             subfield_value.append(item)
