+++ model_card_toolkit/core.py
@@ -17,20 +17,27 @@ The Model Card Toolkit (MCT) provides a set of utilities to generate Model Cards
 from trained models, evaluations, and datasets in ML pipelines.
 """
 
+import logging
 import os
 import pkgutil
 import tempfile
 from typing import Any, Dict, Optional, Union
 
 import jinja2
-import tensorflow_model_analysis as tfma
-from absl import logging
 
+from model_card_toolkit import dependencies
 from model_card_toolkit.model_card import ModelCard
 from model_card_toolkit.proto import model_card_pb2
-from model_card_toolkit.utils import graphics
-from model_card_toolkit.utils import source as src
-from model_card_toolkit.utils import tfx_util
+
+# Imports that require optional dependencies.
+try:
+  import tensorflow_model_analysis as tfma  # pylint: disable=ungrouped-imports
+
+  from model_card_toolkit.utils import tf_graphics, tf_utils
+  from model_card_toolkit.utils.tf_sources import MlmdSource, Source
+except ImportError:
+  MlmdSource = None
+  Source = None
 
 # Constants about provided UI templates.
 _UI_TEMPLATES = (
@@ -90,8 +97,8 @@ class ModelCardToolkit():
   def __init__(
       self,
       output_dir: Optional[str] = None,
-      mlmd_source: Optional[src.MlmdSource] = None,
-      source: Optional[src.Source] = None,
+      mlmd_source: Optional[MlmdSource] = None,
+      source: Optional[Source] = None,
   ):
     """Initializes the ModelCardToolkit.
 
@@ -112,7 +119,12 @@ class ModelCardToolkit():
 
     Raises:
       ValueError: If a model cannot be found at mlmd_source.model_uri.
+      ImportError: If a `source` or `mlmd_source` is provided but necessary
+      extra dependencies are not installed.
     """
+
+    if source or mlmd_source:
+      dependencies.assert_tensorflow_extra_deps_installed()
 
     self.output_dir = output_dir or tempfile.mkdtemp()
     self._mcta_proto_file = os.path.join(self.output_dir, _MCTA_PROTO_FILE)
@@ -129,7 +141,7 @@ class ModelCardToolkit():
     if mlmd_source:
       self._process_mlmd_source(mlmd_source)
 
-  def _process_mlmd_source(self, mlmd_source: src.MlmdSource) -> None:
+  def _process_mlmd_source(self, mlmd_source: MlmdSource) -> None:
     """Process the MLMD source.
 
     This gets the MLMD store, and the artifact corresponding to model_uri.
@@ -213,23 +225,23 @@ class ModelCardToolkit():
               self._source.tfma.metrics_include
               or self._source.tfma.metrics_exclude
           ):
-            eval_result = tfx_util.filter_metrics(
+            eval_result = tf_utils.filter_metrics(
                 eval_result, self._source.tfma.metrics_include,
                 self._source.tfma.metrics_exclude
             )
-          tfx_util.annotate_eval_result_metrics(model_card, eval_result)
-          graphics.annotate_eval_result_plots(model_card, eval_result)
+          tf_utils.annotate_eval_result_metrics(model_card, eval_result)
+          tf_graphics.annotate_eval_result_plots(model_card, eval_result)
         else:
           logging.info('EvalResult not found at path %s', eval_result_path)
     if self._store:
-      metrics_artifacts = tfx_util.get_metrics_artifacts_for_model(
+      metrics_artifacts = tf_utils.get_metrics_artifacts_for_model(
           self._store, self._artifact_with_model_uri.id
       )
       for metrics_artifact in metrics_artifacts:
-        eval_result = tfx_util.read_metrics_eval_result(metrics_artifact.uri)
+        eval_result = tf_utils.read_metrics_eval_result(metrics_artifact.uri)
         if eval_result is not None:
-          tfx_util.annotate_eval_result_metrics(model_card, eval_result)
-          graphics.annotate_eval_result_plots(model_card, eval_result)
+          tf_utils.annotate_eval_result_metrics(model_card, eval_result)
+          tf_graphics.annotate_eval_result_plots(model_card, eval_result)
     return model_card
 
   def _annotate_dataset_statistics(self, model_card: ModelCard) -> ModelCard:
@@ -255,22 +267,22 @@ class ModelCardToolkit():
             self._source.tfdv.features_include
             or self._source.tfdv.features_exclude
         ):
-          data_stats = tfx_util.read_stats_protos_and_filter_features(
+          data_stats = tf_utils.read_stats_protos_and_filter_features(
               dataset_stats_path, self._source.tfdv.features_include,
               self._source.tfdv.features_exclude
           )
         else:
-          data_stats = tfx_util.read_stats_protos(dataset_stats_path)
-        graphics.annotate_dataset_feature_statistics_plots(
+          data_stats = tf_utils.read_stats_protos(dataset_stats_path)
+        tf_graphics.annotate_dataset_feature_statistics_plots(
             model_card, data_stats
         )
     if self._store:
-      stats_artifacts = tfx_util.get_stats_artifacts_for_model(
+      stats_artifacts = tf_utils.get_stats_artifacts_for_model(
           self._store, self._artifact_with_model_uri.id
       )
       for stats_artifact in stats_artifacts:
-        data_stats = tfx_util.read_stats_protos(stats_artifact.uri)
-        graphics.annotate_dataset_feature_statistics_plots(
+        data_stats = tf_utils.read_stats_protos(stats_artifact.uri)
+        tf_graphics.annotate_dataset_feature_statistics_plots(
             model_card, data_stats
         )
     return model_card
@@ -297,7 +309,7 @@ class ModelCardToolkit():
 
     If Source is provided, pre-populate ModelCard fields with data from Source.
     If MLMD store is provided, pre-populate ModelCard fields with data from
-    MLMD. See `model_card_toolkit.utils.tfx_util` and
+    MLMD. See `model_card_toolkit.utils.tf_utils` and
     `model_card_toolkit.utils.graphics` documentation for more details.
 
     Returns:
@@ -305,7 +317,7 @@ class ModelCardToolkit():
     """
     # Pre-populate ModelCard fields
     if self._store:
-      model_card = tfx_util.generate_model_card_for_model(
+      model_card = tf_utils.generate_model_card_for_model(
           self._store, self._artifact_with_model_uri.id
       )
     else:
@@ -436,6 +448,6 @@ class ModelCardToolkit():
     )
 
     # Write the model card document file and return its contents.
-    mode_card_file_path = os.path.join(self._model_cards_dir, output_file)
-    self._write_file(mode_card_file_path, model_card_file_content)
+    model_card_file_path = os.path.join(self._model_cards_dir, output_file)
+    self._write_file(model_card_file_path, model_card_file_content)
     return model_card_file_content
