+++ model_card_toolkit/utils/tf_utils.py
